requirements.txt
py-pure-client==1.62.0
ldap3==2.9.1
PySocks==1.7.1